+++ app/src/androidTest/java/woowacourse/payments/ui/newcard/NewCardContentsTest.kt
@@ -1,6 +1,5 @@
 package woowacourse.payments.ui.newcard
 
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.test.SemanticsNodeInteraction
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
@@ -18,7 +17,7 @@ class NewCardContentsTest {
     @Test
     fun 카드_번호가_16자_미만이면_경고_메시지가_표시된다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("카드 번호")
 
         // when
@@ -31,7 +30,7 @@ class NewCardContentsTest {
     @Test
     fun 카드_번호를_16자_넘게_입력할_경우_첫_16자만_표시된다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("카드 번호")
 
         // when
@@ -44,7 +43,7 @@ class NewCardContentsTest {
     @Test
     fun 카드_번호에_숫자가_아닌_문자가_있으면_경고_메시지가_표시된다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("카드 번호")
 
         // when
@@ -57,7 +56,7 @@ class NewCardContentsTest {
     @Test
     fun 카드_번호를_입력_시_4자_단위로_기호를_삽입한다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("카드 번호")
 
         // when
@@ -70,7 +69,7 @@ class NewCardContentsTest {
     @Test
     fun 만료일을_4자_넘게_입력할_경우_첫_4자만_표시된다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("만료일")
 
         // when
@@ -83,7 +82,7 @@ class NewCardContentsTest {
     @Test
     fun 만료일이_MM_YY_형식이_아니면_경고_메시지가_표시된다_1() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("만료일")
 
         // when
@@ -96,7 +95,7 @@ class NewCardContentsTest {
     @Test
     fun 만료일이_MM_YY_형식이_아니면_경고_메시지가_표시된다_2() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("만료일")
 
         // when
@@ -109,7 +108,7 @@ class NewCardContentsTest {
     @Test
     fun 만료일이_이미_지났으면_경고_메시지가_표시된다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current, YearMonth.of(2025, 9)) }
+        composeTestRule.setContent { NewCardContents(YearMonth.of(2025, 9)) }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("만료일")
 
         // when
@@ -122,7 +121,7 @@ class NewCardContentsTest {
     @Test
     fun 만료일을_입력_시_2자_단위로_기호를_삽입한다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("만료일")
 
         // when
@@ -135,7 +134,7 @@ class NewCardContentsTest {
     @Test
     fun 카드_소유자_이름을_30자_넘게_입력할_경우_첫_30자만_표시된다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("카드 소유자 이름(선택)")
 
         // when
@@ -148,7 +147,7 @@ class NewCardContentsTest {
     @Test
     fun 카드_소유자_이름에_입력된_문자의_개수를_표시한다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("카드 소유자 이름(선택)")
 
         // when
@@ -161,7 +160,7 @@ class NewCardContentsTest {
     @Test
     fun 비밀가_4자_미만이면_경고_메시지가_표시된다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("비밀번호")
 
         // when
@@ -174,7 +173,7 @@ class NewCardContentsTest {
     @Test
     fun 비밀번호를_4자_넘게_입력할_경우_첫_4자만_표시된다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("비밀번호")
 
         // when
@@ -187,7 +186,7 @@ class NewCardContentsTest {
     @Test
     fun 비밀번호에_숫자가_아닌_문자가_있으면_경고_메시지가_표시된다() {
         // given
-        composeTestRule.setContent { NewCardContents(LocalContext.current) }
+        composeTestRule.setContent { NewCardContents() }
         val target: SemanticsNodeInteraction = composeTestRule.onNodeWithText("비밀번호")
 
         // when
